feat(worldgen): introduce populateTownNpcs and integrate into town_gen; wire in TownAI NPC population.
