+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/worldgen/town_gen.js
@@ -27,7 +27,7 @@ import { getTownBuildingConfig, getInnSizeConfig, getCastleKeepSizeConfig, getTo
 import { buildBaseTown, buildPlaza, carveBuildingRect } from "./town/layout_core.js";
 import { buildOutdoorMask, repairBuildingPerimeters, placeWindowsOnAll } from "./town/windows.js";
 import { addProp, addSignNear, addShopSignInside, dedupeShopSigns, dedupeWelcomeSign, cleanupDanglingProps } from "./town/signs.js";
-import { spawnGateGreeters, enforceGateNPCLimit } from "./town/npcs_bootstrap.js";
+import { spawnGateGreeters, enforceGateNPCLimit, populateTownNpcs } from "./town/npcs_bootstrap.js";
 
 function inBounds(ctx, x, y) {
   try {
@@ -1863,181 +1863,8 @@ function generate(ctx) {
   // Repair pass: enforce solid building perimeters (convert any non-door/window on borders to WALL)
   repairBuildingPerimeters(ctx, buildings);
 
-  // NPCs via TownAI if present
-  ctx.npcs = [];
-  try {
-    if (ctx && ctx.TownAI && typeof ctx.TownAI.populateTown === "function") {
-      ctx.TownAI.populateTown(ctx);
-    } else {
-      const TAI = ctx.TownAI || getMod(ctx, "TownAI");
-      if (TAI && typeof TAI.populateTown === "function") {
-        TAI.populateTown(ctx);
-      }
-    }
-  } catch (_) {}
-
-  // One special cat: Jekku (spawn in the designated town only)
-// Another special cat: Pulla (same behavior, different name, spawns in its designated town)
-  (function placeSpecialCats() {
-    try {
-      const wx = (ctx.worldReturnPos && typeof ctx.worldReturnPos.x === "number") ? ctx.worldReturnPos.x : ctx.player.x;
-      const wy = (ctx.worldReturnPos && typeof ctx.worldReturnPos.y === "number") ? ctx.worldReturnPos.y : ctx.player.y;
-      const info = (ctx.world && Array.isArray(ctx.world.towns)) ? ctx.world.towns.find(t => t.x === wx && t.y === wy) : null;
-      if (!info) return;
-
-      function spawnCatOnce(nameCheck, displayName) {
-        // Avoid duplicate by name if already present
-        if (Array.isArray(ctx.npcs) && ctx.npcs.some(n => String(n.name || "").toLowerCase() === nameCheck)) return;
-        // Prefer a free floor near the plaza
-        const spots = [
-          { x: ctx.townPlaza.x + 1, y: ctx.townPlaza.y },
-          { x: ctx.townPlaza.x - 1, y: ctx.townPlaza.y },
-          { x: ctx.townPlaza.x, y: ctx.townPlaza.y + 1 },
-          { x: ctx.townPlaza.x, y: ctx.townPlaza.y - 1 },
-          { x: ctx.townPlaza.x + 2, y: ctx.townPlaza.y },
-          { x: ctx.townPlaza.x - 2, y: ctx.townPlaza.y },
-          { x: ctx.townPlaza.x, y: ctx.townPlaza.y + 2 },
-          { x: ctx.townPlaza.x, y: ctx.townPlaza.y - 2 },
-        ];
-        let pos = null;
-        for (const s of spots) { if (_isFreeTownFloor(ctx, s.x, s.y)) { pos = s; break; } }
-        if (!pos) {
-          // Fallback: any free floor near plaza
-          for (let oy = -3; oy <= 3 && !pos; oy++) {
-            for (let ox = -3; ox <= 3 && !pos; ox++) {
-              const x = ctx.townPlaza.x + ox, y = ctx.townPlaza.y + oy;
-              if (_isFreeTownFloor(ctx, x, y)) pos = { x, y };
-            }
-          }
-        }
-        if (!pos) pos = { x: ctx.townPlaza.x, y: ctx.townPlaza.y };
-        ctx.npcs.push({ x: pos.x, y: pos.y, name: displayName, kind: "cat", lines: ["Meow.", "Purr."], pet: true });
-      }
-
-      if (info.jekkuHome) {
-        spawnCatOnce("jekku", "Jekku");
-      }
-      if (info.pullaHome) {
-        spawnCatOnce("pulla", "Pulla");
-      }
-    } catch (_) {}
-  })();
-
-  // Roaming villagers near plaza (some promoted to town guards)
-  const GD8 = getGameData(ctx);
-  const ND = (GD8 && GD8.npcs) ? GD8.npcs : null;
-  const baseLines = (ND && Array.isArray(ND.residentLines) && ND.residentLines.length)
-    ? ND.residentLines
-    : [
-        "Rest your feet a while.",
-        "The dungeon is dangerous.",
-        "Buy supplies before you go.",
-        "Lovely day on the plaza.",
-        "Care for a drink at the well?"
-      ];
-  const lines = [
-    `Welcome to ${ctx.townName || "our town"}.`,
-    ...baseLines
-  ];
-  const guardLines = (ND && Array.isArray(ND.guardLines) && ND.guardLines.length)
-    ? ND.guardLines
-    : [
-        "Stay out of trouble.",
-        "We keep the town safe.",
-        "Eyes open, blade sharp.",
-        "The gate is watched day and night."
-      ];
-  const tbCount = Array.isArray(ctx.townBuildings) ? ctx.townBuildings.length : 12;
-  // Dedicated guard barracks building (if present) for guard homes.
-  const guardBarracks = Array.isArray(ctx.townBuildings)
-    ? ctx.townBuildings.find(b => b && b.prefabId && String(b.prefabId).toLowerCase().includes("guard_barracks"))
-    : null;
-
-  // Population targets (roamers/guards) driven by town.json when present.
-  const popTargets = getTownPopulationTargets(TOWNCFG, townSize, townKind, tbCount);
-  const roamTarget = popTargets.roamTarget;
-  let guardTarget = popTargets.guardTarget;
-
-  let placed = 0, placedGuards = 0, tries = 0;
-  while (placed < roamTarget && tries++ < 800) {
-    const onRoad = ctx.rng() < 0.4;
-    let x, y;
-    if (onRoad) {
-      if (ctx.rng() < 0.5) { y = gate.y; x = Math.max(2, Math.min(W - 3, Math.floor(ctx.rng() * (W - 4)) + 2)); }
-      else { x = plaza.x; y = Math.max(2, Math.min(H - 3, Math.floor(ctx.rng() * (H - 4)) + 2)); }
-    } else {
-      const ox = Math.floor(ctx.rng() * 21) - 10;
-      const oy = Math.floor(ctx.rng() * 17) - 8;
-      x = Math.max(1, Math.min(W - 2, plaza.x + ox));
-      y = Math.max(1, Math.min(H - 2, plaza.y + oy));
-    }
-    if (ctx.map[y][x] !== ctx.TILES.FLOOR && ctx.map[y][x] !== ctx.TILES.DOOR && ctx.map[y][x] !== ctx.TILES.ROAD) continue;
-    if (x === ctx.player.x && y === ctx.player.y) continue;
-    if (_manhattan(ctx, ctx.player.x, ctx.player.y, x, y) <= 1) continue;
-    if (ctx.npcs.some(n => n.x === x && n.y === y)) continue;
-    if (ctx.townProps.some(p => p.x === x && p.y === y)) continue;
-
-    // Prefer to turn road/near-gate roamers into guards, up to guardTarget
-    const nearGate = _manhattan(ctx, x, y, gate.x, gate.y) <= 6;
-    const canBeGuard = placedGuards < guardTarget && (onRoad || nearGate || ctx.rng() < 0.25);
-
-    // Assign a home immediately to avoid "no-home" diagnostics for roamers/guards
-    let homeRef = null;
-    try {
-      const tbs = Array.isArray(ctx.townBuildings) ? ctx.townBuildings : [];
-      if (tbs.length) {
-        let b = null;
-        if (canBeGuard && guardBarracks) {
-          b = guardBarracks;
-        } else {
-          b = tbs[Math.floor(rng() * tbs.length)];
-        }
-        if (b) {
-          const hx = Math.max(b.x + 1, Math.min(b.x + b.w - 2, (b.x + ((b.w / 2) | 0))));
-          const hy = Math.max(b.y + 1, Math.min(b.y + b.h - 2, (b.y + ((b.h / 2) | 0))));
-          const door = (b && b.door && typeof b.door.x === "number" && typeof b.door.y === "number") ? { x: b.door.x, y: b.door.y } : null;
-          homeRef = { building: b, x: hx, y: hy, door };
-        }
-      }
-    } catch (_) {}
-
-    const likesInn = rng() < 0.45;
-    if (canBeGuard) {
-      const guardIndex = placedGuards + 1;
-      const eliteChance = 0.3;
-      const isEliteGuard = (ctx && typeof ctx.rng === "function") ? (ctx.rng() < eliteChance) : (Math.random() < eliteChance);
-      const guardType = isEliteGuard ? "guard_elite" : "guard";
-      const guardName = isEliteGuard ? `Guard captain ${guardIndex}` : `Guard ${guardIndex}`;
-      const baseHp = isEliteGuard ? 28 : 22;
-      const hpJitter = (ctx && typeof ctx.rng === "function") ? Math.floor(ctx.rng() * 6) : 0;
-      const hp = baseHp + hpJitter;
-      ctx.npcs.push({
-        x,
-        y,
-        name: guardName,
-        lines: guardLines,
-        isGuard: true,
-        guard: true,
-        guardType,
-        hp,
-        maxHp: hp,
-        _guardPost: { x, y },
-        _likesInn: likesInn,
-        _home: homeRef
-      });
-      placedGuards++;
-    } else {
-      ctx.npcs.push({
-        x,
-        y,
-        name: `Villager ${placed + 1}`,
-        lines,
-        _likesInn: likesInn,
-        _home: homeRef
-      });
-    }
-    placed++;
-  }
+  // NPCs via TownAI + special cats + roaming villagers/guards
+  populateTownNpcs(ctx, W, H, gate, plaza, townSize, townKind, TOWNCFG, info, rng);
 
   // Visibility reset for town
   // Start unseen; player FOV will reveal tiles and mark memory.
